+++ test/test_helper.rb
@@ -6,9 +6,7 @@ require 'active_record'
 require 'active_support'
 require 'active_support/test_case'
 
-require 'lib/association_with_operator'
 require 'lib/record_with_operator'
-require 'init'
 
 config = YAML::load(IO.read(File.dirname(__FILE__) + '/database.yml'))
 ActiveRecord::Base.logger = Logger.new(File.dirname(__FILE__) + "/debug.log")
